Package the module and prepare for initial upload to pypi
Includes:
* Changes to directory structure
* Changes to config files and metadata
* Minor improvements in error handling
* In the Main.py test script, Radio Serial Number is now read from an env variable
