--- FlexModule/setup.py
@@ -1,11 +1,0 @@
-from setuptools import setup
-
-setup(name='FlexModule',
-	version='0.1',
-	description='Python API for the FLEX 6000/6400 radio',
-	url='',
-	author='Jonny Slim',
-	author_email='slimjm@cardiff.ac.uk',
-	license='GPL',
-	packages=['FlexModule'],
-	zip_safe=False)
